requirements.txt
kafka-python
boto3
requests